requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
slack-sdk
boto3
requests
certifi
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
cachetools
rsa
pyasn1
pyasn1-modules
httplib2
uritemplate
langchain-google-genai
langchain-community
